BUG: fix SuiteSparse_long on windows
Fundamentally, SuiteSparse requires a 64 bit
integer to operate correctly. On GCC, this type
is provided by the "long" type definition.
However, although the "long" type happens to meet
this criteria, is not necesarrily the case that
a "long" is 64 bits because it is officially
defined to be between 32 and 64 bits.

Because MSVC implemented "long" differently from
GCC, although both complied with the specification,
MSVC failed when GCC did not.

To fix this error, the SWIG file was updated to use
the C99 int64_t, which is standardized at 64 bits.
A compatibility header was added for MSVC, which
does not fully support C99 (the compatibility
header can be removed sometime in the future).

+++ scikits/umfpack/umfpack.py
@@ -505,13 +505,16 @@ class UmfpackContext(Struct):
         if self.isReal:
             status, self._symbolic\
                     = self.funs.symbolic(mtx.shape[0], mtx.shape[1],
-                                          mtx.indptr, indx, mtx.data,
+                                          mtx.indptr.astype('int32'),
+                                          indx.astype('int32'),
+                                          mtx.data,
                                           self.control, self.info)
         else:
             real, imag = mtx.data.real.copy(), mtx.data.imag.copy()
             status, self._symbolic\
                     = self.funs.symbolic(mtx.shape[0], mtx.shape[1],
-                                          mtx.indptr, indx,
+                                          mtx.indptr.astype('int32'),
+                                          indx.astype('int32'),
                                           real, imag,
                                           self.control, self.info)
